Completed tests of all parameters. Working fine. NOTE THAT a hotel's multiple string values for nearby sites and nearby day tours have been changed to a single nearby site and single nearby tour. We may change this back to multiple values later, but right now it's just adding a layer of complexity we don't really need, in return for value that's not required for our program.
Also note that in SearchView there is a test...() function which contains code for creating a parameter list. You can use that code to create your own parameter list if you want to call the HotelManager.search function
with the overload parameter list, rather than just with dates and guest numbers.